+++ Modules/Intent.Modules.Constants/Designers.cs
@@ -4,4 +4,6 @@ public static class Designers
 {
     public const string Domain = "Domain";
     public const string Services = "Services";
+
+    public const string VisualStudioId = "0701433c-36c0-4569-b1f4-9204986b587d";
 }