+++ Modules/Intent.Modules.Constants/ProjectStereotypes.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Intent.Modules.Constants;
+public static class ProjectStereotypes
+{
+    public static class LaunchSettings
+    {
+        public const string Id = "fd52e71a-3810-4ae7-ae40-4e1514903d25";
+        public const string BaseUrlId = "1b85d04d-e80a-4640-9fee-558711473fa8";
+    }
+}