+++ Modules/Intent.Modules.Integration.HttpClients/Templates/HttpClientConfiguration/HttpClientConfigurationTemplatePartial.cs
@@ -1,6 +1,3 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
 using Intent.Engine;
 using Intent.Metadata.Models;
 using Intent.Modules.Application.Contracts.Clients.Templates;
@@ -12,6 +9,7 @@ using Intent.Modules.Common.CSharp.DependencyInjection;
 using Intent.Modules.Common.CSharp.Templates;
 using Intent.Modules.Common.Templates;
 using Intent.Modules.Common.VisualStudio;
+using Intent.Modules.Constants;
 using Intent.Modules.Contracts.Clients.Shared;
 using Intent.Modules.Integration.HttpClients.Settings;
 using Intent.Modules.Integration.HttpClients.Shared.Templates;
@@ -19,8 +17,13 @@ using Intent.Modules.Integration.HttpClients.Shared.Templates.Adapters;
 using Intent.Modules.Integration.HttpClients.Shared.Templates.HttpClientConfiguration;
 using Intent.Modules.Integration.HttpClients.Templates.HttpClient;
 using Intent.Modules.Metadata.WebApi.Models;
+using Intent.Persistence;
 using Intent.RoslynWeaver.Attributes;
 using Intent.Templates;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading;
 
 [assembly: DefaultIntentManaged(Mode.Fully)]
 [assembly: IntentTemplate("Intent.ModuleBuilder.CSharp.Templates.CSharpTemplatePartial", Version = "1.0")]
@@ -54,7 +57,6 @@ namespace Intent.Modules.Integration.HttpClients.Templates.HttpClientConfigurati
         {
             base.BeforeTemplateExecution();
 
-
             if (ExecutionContext.Settings.GetIntegrationHttpClientSettings().AuthorizationSetup().IsClientAccessTokenManagement())
             {
                 var dotNetVersion = OutputTarget.GetMaxNetAppVersion();
@@ -72,17 +74,16 @@ namespace Intent.Modules.Integration.HttpClients.Templates.HttpClientConfigurati
                 ExecutionContext.EventDispatcher.Publish(new AppSettingRegistrationRequest("IdentityClients:default:Scope", "api"));
             }
 
-
             var proxies = _typedModels.DistinctBy(x => x.Id);
-            var groups = proxies.Select(p => GetGroupName(p.InternalElement)).Distinct();
-            foreach (var groupName in groups)
+            var proxySettings = proxies.Select(p => (GroupName: GetGroupName(p.InternalElement), Url: ProxyUrlHelper.GetProxyApplicationtUrl(p.InternalElement, ExecutionContext))).DistinctBy(d => d.GroupName);
+            foreach (var (GroupName, Url) in proxySettings)
             {
-                ExecutionContext.EventDispatcher.Publish(new AppSettingRegistrationRequest(GetConfigKey(groupName, "Uri"), "https://localhost:{app_port}/"));
+                ExecutionContext.EventDispatcher.Publish(new AppSettingRegistrationRequest(GetConfigKey(GroupName, "Uri"), Url));
                 if (ExecutionContext.Settings.GetIntegrationHttpClientSettings().AuthorizationSetup().IsClientAccessTokenManagement())
                 {
-                    ExecutionContext.EventDispatcher.Publish(new AppSettingRegistrationRequest(GetConfigKey(groupName, "IdentityClientKey"), "default"));
+                    ExecutionContext.EventDispatcher.Publish(new AppSettingRegistrationRequest(GetConfigKey(GroupName, "IdentityClientKey"), "default"));
                 }
-                ExecutionContext.EventDispatcher.Publish(new AppSettingRegistrationRequest(GetConfigKey(groupName, "Timeout"), "00:01:00"));
+                ExecutionContext.EventDispatcher.Publish(new AppSettingRegistrationRequest(GetConfigKey(GroupName, "Timeout"), "00:01:00"));
 
             }
         }