+++ Modules/Intent.Modules.AspNetCore.IntegrationTesting/release-notes.md
@@ -2,6 +2,7 @@
 
 - Improvement: Updated NuGet package versions.
 - Improvement: adding using clause for `HttpRequestMessage` in `HttpCLient`s.
+- Improvement: Updated referenced projects.
  
 ### Version 2.0.9