+++ Modules/Intent.Modules.Common.UnitOfWork/release-notes.md
@@ -1,3 +1,7 @@
+### Version 1.0.1
+
+- Fixed: Reference no longer added to the VS designer
+
 ### Version 1.0.0
 
 Initial Release